+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - fetching of Good To Go custom fields
+- snapshot image uploading for Arc GIS form element
 
 ## [23.1.1] - 2025-07-25